task-host/requirements.txt
fastapi[all]>=0.75
uvicorn[standard]>=0.15
pydantic
Pillow

# For static images
aiofiles